@@ -5,6 +5,7 @@ import 'package:flutter_portal/flutter_portal.dart';
 import 'package:meta/meta.dart';
 import 'package:stream_chat_flutter/conditional_parent_builder/conditional_parent_builder.dart';
 import 'package:stream_chat_flutter/platform_widget_builder/platform_widget_builder.dart';
+import 'package:stream_chat_flutter/src/attachment/builder/attachment_widget_builder.dart';
 import 'package:stream_chat_flutter/src/context_menu_items/context_menu_reaction_picker.dart';
 import 'package:stream_chat_flutter/src/context_menu_items/stream_chat_context_menu_item.dart';
 import 'package:stream_chat_flutter/src/dialogs/dialogs.dart';
@@ -41,28 +42,22 @@ enum DisplayWidget {
 /// {@endtemplate}
 class StreamMessageWidget extends StatefulWidget {
   /// {@macro messageWidget}
-  StreamMessageWidget({
+  const StreamMessageWidget({
     super.key,
     required this.message,
     required this.messageTheme,
     this.reverse = false,
     this.translateUserAvatar = true,
     this.shape,
-    this.attachmentShape,
     this.borderSide,
-    this.attachmentBorderSide,
     this.borderRadiusGeometry,
-    this.attachmentBorderRadiusGeometry,
+    this.attachmentShape,
     this.onMentionTap,
     this.onMessageTap,
     this.onReactionsTap,
     this.onReactionsHover,
-    bool? showReactionPicker,
-    @Deprecated('Use `showReactionPicker` instead')
-    bool showReactionPickerIndicator = true,
-    @internal
-    @Deprecated('Use `showReactionPicker` instead')
-    this.showReactionPickerTail,
+    this.showReactionPicker = true,
+    @internal this.showReactionPickerTail = false,
     this.showUserAvatar = DisplayWidget.show,
     this.showSendingIndicator = true,
     this.showThreadReplyIndicator = false,
@@ -90,16 +85,8 @@ class StreamMessageWidget extends StatefulWidget {
     this.quotedMessageBuilder,
     this.editMessageInputBuilder,
     this.textBuilder,
-    @Deprecated('''
-    Use [bottomRowBuilderWithDefaultWidget] instead.
-    Will be removed in the next major version.
-    ''') this.bottomRowBuilder,
     this.bottomRowBuilderWithDefaultWidget,
-    @Deprecated('''
-    Use [bottomRowBuilderWithDefaultWidget] instead.
-    Will be removed in the next major version.
-    ''') this.deletedBottomRowBuilder,
-    this.customAttachmentBuilders,
+    this.attachmentBuilders,
     this.padding,
     this.textPadding = const EdgeInsets.symmetric(
       horizontal: 16,
@@ -110,198 +97,161 @@ class StreamMessageWidget extends StatefulWidget {
     this.onQuotedMessageTap,
     this.customActions = const [],
     this.onAttachmentTap,
-    @Deprecated('''
-    Use [bottomRowBuilderWithDefaultWidget] instead.
-    Will be removed in the next major version.
-    ''') this.usernameBuilder,
     this.imageAttachmentThumbnailSize = const Size(400, 400),
     this.imageAttachmentThumbnailResizeType = 'clip',
     this.imageAttachmentThumbnailCropType = 'center',
     this.attachmentActionsModalBuilder,
-  })  : assert(
-          bottomRowBuilder == null || bottomRowBuilderWithDefaultWidget == null,
-          'You can only use one of the two bottom row builders',
-        ),
-        showReactionPicker = showReactionPicker ?? showReactionPickerIndicator,
-        attachmentBuilders = {
-          'image': (context, message, attachments) {
-            final border = RoundedRectangleBorder(
-              side: attachmentBorderSide ??
-                  BorderSide(
-                    color: StreamChatTheme.of(context).colorTheme.borders,
-                  ),
-              borderRadius: attachmentBorderRadiusGeometry ?? BorderRadius.zero,
-            );
+  });
 
-            final mediaQueryData = MediaQuery.of(context);
-            if (attachments.length > 1) {
-              return Padding(
-                padding: attachmentPadding,
-                child: WrapAttachmentWidget(
-                  attachmentWidget: Material(
-                    color: messageTheme.messageBackgroundColor,
-                    child: StreamImageGroup(
-                      constraints: BoxConstraints(
-                        maxWidth: 400,
-                        minWidth: 400,
-                        maxHeight: mediaQueryData.size.height * 0.3,
-                      ),
-                      images: attachments,
-                      message: message,
-                      messageTheme: messageTheme,
-                      onShowMessage: onShowMessage,
-                      onReplyMessage: onReplyTap,
-                      onAttachmentTap: onAttachmentTap,
-                      imageThumbnailSize: imageAttachmentThumbnailSize,
-                      imageThumbnailResizeType:
-                          imageAttachmentThumbnailResizeType,
-                      imageThumbnailCropType: imageAttachmentThumbnailCropType,
-                      attachmentActionsModalBuilder:
-                          attachmentActionsModalBuilder,
-                    ),
-                  ),
-                  attachmentShape: border,
-                ),
-              );
-            }
-
-            return WrapAttachmentWidget(
-              attachmentWidget: StreamImageAttachment(
-                attachment: attachments[0],
-                message: message,
-                messageTheme: messageTheme,
-                constraints: BoxConstraints(
-                  maxWidth: 400,
-                  minWidth: 400,
-                  maxHeight: mediaQueryData.size.height * 0.3,
-                ),
-                onShowMessage: onShowMessage,
-                onReplyMessage: onReplyTap,
-                onAttachmentTap: onAttachmentTap != null
-                    ? () {
-                        onAttachmentTap.call(message, attachments[0]);
-                      }
-                    : null,
-                imageThumbnailSize: imageAttachmentThumbnailSize,
-                imageThumbnailResizeType: imageAttachmentThumbnailResizeType,
-                imageThumbnailCropType: imageAttachmentThumbnailCropType,
-                attachmentActionsModalBuilder: attachmentActionsModalBuilder,
-              ),
-              attachmentShape: border,
-            );
-          },
-          'video': (context, message, attachments) {
-            final border = RoundedRectangleBorder(
-              side: attachmentBorderSide ??
-                  BorderSide(
-                    color: StreamChatTheme.of(context).colorTheme.borders,
-                  ),
-              borderRadius: attachmentBorderRadiusGeometry ?? BorderRadius.zero,
-            );
-
-            return WrapAttachmentWidget(
-              attachmentWidget: Column(
-                children: attachments.map((attachment) {
-                  final mediaQueryData = MediaQuery.of(context);
-                  return StreamVideoAttachment(
-                    attachment: attachment,
-                    messageTheme: messageTheme,
-                    constraints: BoxConstraints(
-                      maxWidth: 400,
-                      minWidth: 400,
-                      maxHeight: mediaQueryData.size.height * 0.3,
-                    ),
-                    message: message,
-                    onShowMessage: onShowMessage,
-                    onReplyMessage: onReplyTap,
-                    onAttachmentTap: onAttachmentTap != null
-                        ? () {
-                            onAttachmentTap(message, attachment);
-                          }
-                        : null,
-                    attachmentActionsModalBuilder:
-                        attachmentActionsModalBuilder,
-                  );
-                }).toList(),
-              ),
-              attachmentShape: border,
-            );
-          },
-          'giphy': (context, message, attachments) {
-            final attachmentWidget = Column(
-              children: [
-                ...attachments.map((attachment) {
-                  final mediaQueryData = MediaQuery.of(context);
-                  return StreamGiphyAttachment(
-                    attachment: attachment,
-                    message: message,
-                    constraints: BoxConstraints(
-                      maxWidth: 400,
-                      minWidth: 400,
-                      maxHeight: mediaQueryData.size.height * 0.3,
-                    ),
-                    onShowMessage: onShowMessage,
-                    onReplyMessage: onReplyTap,
-                    onAttachmentTap: onAttachmentTap != null
-                        ? () => onAttachmentTap(message, attachment)
-                        : null,
-                  );
-                }),
-              ],
-            );
-
-            // If the message is ephemeral, we don't want to show the border.
-            if (message.isEphemeral) return attachmentWidget;
-
-            final color = StreamChatTheme.of(context).colorTheme.borders;
-            final border = RoundedRectangleBorder(
-              side: attachmentBorderSide ?? BorderSide(color: color),
-              borderRadius: attachmentBorderRadiusGeometry ?? BorderRadius.zero,
-            );
-
-            return WrapAttachmentWidget(
-              attachmentShape: border,
-              attachmentWidget: attachmentWidget,
-            );
-          },
-          'file': (context, message, attachments) {
-            final border = RoundedRectangleBorder(
-              side: attachmentBorderSide ??
-                  BorderSide(
-                    color: StreamChatTheme.of(context).colorTheme.borders,
-                  ),
-              borderRadius: attachmentBorderRadiusGeometry ?? BorderRadius.zero,
-            );
-
-            return Column(
-              children: attachments
-                  .map<Widget>((attachment) {
-                    final mediaQueryData = MediaQuery.of(context);
-                    return WrapAttachmentWidget(
-                      attachmentWidget: StreamFileAttachment(
-                        message: message,
-                        attachment: attachment,
-                        constraints: BoxConstraints(
-                          maxWidth: 400,
-                          minWidth: 400,
-                          maxHeight: mediaQueryData.size.height * 0.3,
-                        ),
-                        onAttachmentTap: onAttachmentTap != null
-                            ? () {
-                                onAttachmentTap(message, attachment);
-                              }
-                            : null,
-                      ),
-                      attachmentShape: border,
-                    );
-                  })
-                  .insertBetween(SizedBox(
-                    height: attachmentPadding.vertical / 2,
-                  ))
-                  .toList(),
-            );
-          },
-        }..addAll(customAttachmentBuilders ?? {});
+  // attachmentBuilders = {
+  //   // Add all default builders
+  //   'image': (context, message, attachments) {
+  //     final color = StreamChatTheme.of(context).colorTheme.borders;
+  //     final border = RoundedRectangleBorder(
+  //       side: attachmentBorderSide ?? BorderSide(color: color),
+  //       borderRadius: attachmentBorderRadiusGeometry ?? BorderRadius.zero,
+  //     );
+  //
+  //     if (attachments.length > 1) {
+  //       return WrapAttachmentWidget(
+  //         attachmentShape: border,
+  //         attachmentWidget: Material(
+  //           color: messageTheme.messageBackgroundColor,
+  //           child: StreamGalleryAttachment(
+  //             constraints: const BoxConstraints.tightFor(
+  //               width: 256,
+  //               height: 195,
+  //             ),
+  //             attachments: attachments,
+  //             message: message,
+  //             itemBuilder: (context, index) {
+  //               return Placeholder();
+  //             },
+  //             // onShowMessage: onShowMessage,
+  //             // onReplyMessage: onReplyTap,
+  //             // onAttachmentTap: onAttachmentTap,
+  //             // imageThumbnailSize: imageAttachmentThumbnailSize,
+  //             // imageThumbnailResizeType:
+  //             //     imageAttachmentThumbnailResizeType,
+  //             // imageThumbnailCropType: imageAttachmentThumbnailCropType,
+  //             // attachmentActionsModalBuilder:
+  //             //     attachmentActionsModalBuilder,
+  //           ),
+  //         ),
+  //       );
+  //     }
+  //
+  //     return WrapAttachmentWidget(
+  //       attachmentShape: border,
+  //       attachmentWidget: StreamImageAttachment(
+  //         message: message,
+  //         image: attachments.first,
+  //         constraints: const BoxConstraints(
+  //           minWidth: 170,
+  //           maxWidth: 256,
+  //           minHeight: 100,
+  //           maxHeight: 300,
+  //         ),
+  //         // onShowMessage: onShowMessage,
+  //         // onReplyMessage: onReplyTap,
+  //         imageThumbnailSize: imageAttachmentThumbnailSize,
+  //         imageThumbnailResizeType: imageAttachmentThumbnailResizeType,
+  //         imageThumbnailCropType: imageAttachmentThumbnailCropType,
+  //         // attachmentActionsModalBuilder: attachmentActionsModalBuilder,
+  //         // onAttachmentTap: onAttachmentTap != null
+  //         //     ? () => onAttachmentTap.call(message, attachments.first)
+  //         //     : null,
+  //       ),
+  //     );
+  //   },
+  //   'video': (context, message, attachments) {
+  //     final color = StreamChatTheme.of(context).colorTheme.borders;
+  //     final border = RoundedRectangleBorder(
+  //       side: attachmentBorderSide ?? BorderSide(color: color),
+  //       borderRadius: attachmentBorderRadiusGeometry ?? BorderRadius.zero,
+  //     );
+  //
+  //     return WrapAttachmentWidget(
+  //       attachmentShape: border,
+  //       attachmentWidget: Column(
+  //         children: [
+  //           ...attachments.map((attachment) {
+  //             return StreamVideoAttachment(
+  //               video: attachment,
+  //               constraints: const BoxConstraints.tightFor(
+  //                 width: 256,
+  //                 height: 195,
+  //               ),
+  //               message: message,
+  //             );
+  //           }),
+  //         ],
+  //       ),
+  //     );
+  //   },
+  //   'giphy': (context, message, attachments) {
+  //     final color = StreamChatTheme.of(context).colorTheme.borders;
+  //     final border = RoundedRectangleBorder(
+  //       side: attachmentBorderSide ?? BorderSide(color: color),
+  //       borderRadius: attachmentBorderRadiusGeometry ?? BorderRadius.zero,
+  //     );
+  //
+  //     return WrapAttachmentWidget(
+  //       attachmentShape: border,
+  //       attachmentWidget: Column(
+  //         children: [
+  //           ...attachments.map((attachment) {
+  //             return StreamGiphyAttachment(
+  //               giphy: attachment,
+  //               message: message,
+  //               constraints: const BoxConstraints(
+  //                 minWidth: 170,
+  //                 maxWidth: 256,
+  //                 minHeight: 100,
+  //                 maxHeight: 300,
+  //               ),
+  //             );
+  //           }),
+  //         ],
+  //       ),
+  //     );
+  //   },
+  //   'file': (context, message, attachments) {
+  //     final color = StreamChatTheme.of(context).colorTheme.borders;
+  //     final border = RoundedRectangleBorder(
+  //       side: attachmentBorderSide ?? BorderSide(color: color),
+  //       borderRadius: attachmentBorderRadiusGeometry ?? BorderRadius.zero,
+  //     );
+  //
+  //     return Column(
+  //       children: [
+  //         ...attachments.map<Widget>((attachment) {
+  //           final mediaQueryData = MediaQuery.of(context);
+  //           return WrapAttachmentWidget(
+  //             attachmentShape: border,
+  //             attachmentWidget: StreamFileAttachment(
+  //               message: message,
+  //               file: attachment,
+  //               constraints: BoxConstraints(
+  //                 maxWidth: 400,
+  //                 minWidth: 400,
+  //                 maxHeight: mediaQueryData.size.height * 0.3,
+  //               ),
+  //               // onAttachmentTap: onAttachmentTap != null
+  //               //     ? () => onAttachmentTap(message, attachment)
+  //               //     : null,
+  //             ),
+  //           );
+  //         }).insertBetween(
+  //           SizedBox(height: attachmentPadding.vertical / 2),
+  //         ),
+  //       ],
+  //     );
+  //   },
+  //
+  //   // Add all custom builders, overriding the defaults if needed.
+  //   ...?customAttachmentBuilders,
+  // };
 
   /// {@template onMentionTap}
   /// Function called on mention tap
@@ -333,31 +283,16 @@ class StreamMessageWidget extends StatefulWidget {
   /// {@endtemplate}
   final Widget Function(BuildContext, Message)? textBuilder;
 
-  /// {@template usernameBuilder}
-  /// Widget builder for building username
-  /// {@endtemplate}
-  final Widget Function(BuildContext, Message)? usernameBuilder;
-
   /// {@template onMessageActions}
   /// Function called on long press
   /// {@endtemplate}
   final void Function(BuildContext, Message)? onMessageActions;
-
-  /// {@template bottomRowBuilder}
-  /// Widget builder for building a bottom row below the message
-  /// {@endtemplate}
-  final BottomRowBuilder? bottomRowBuilder;
 
   /// {@template bottomRowBuilderWithDefaultWidget}
   /// Widget builder for building a bottom row below the message.
   /// Also contains the default bottom row widget.
   /// {@endtemplate}
   final BottomRowBuilderWithDefaultWidget? bottomRowBuilderWithDefaultWidget;
-
-  /// {@template deletedBottomRowBuilder}
-  /// Widget builder for building a bottom row below a deleted message
-  /// {@endtemplate}
-  final Widget Function(BuildContext, Message)? deletedBottomRowBuilder;
 
   /// {@template userAvatarBuilder}
   /// Widget builder for building user avatar
@@ -399,20 +334,10 @@ class StreamMessageWidget extends StatefulWidget {
   /// {@endtemplate}
   final BorderSide? borderSide;
 
-  /// {@template attachmentBorderSide}
-  /// The borderSide of an attachment
-  /// {@endtemplate}
-  final BorderSide? attachmentBorderSide;
-
   /// {@template borderRadiusGeometry}
   /// The border radius of the message text
   /// {@endtemplate}
   final BorderRadiusGeometry? borderRadiusGeometry;
-
-  /// {@template attachmentBorderRadiusGeometry}
-  /// The border radius of an attachment
-  /// {@endtemplate}
-  final BorderRadiusGeometry? attachmentBorderRadiusGeometry;
 
   /// {@template padding}
   /// The padding of the widget
@@ -474,12 +399,6 @@ class StreamMessageWidget extends StatefulWidget {
   /// Used in [StreamMessageReactionsModal] and [MessageActionsModal].
   /// {@endtemplate}
   final bool showReactionPicker;
-
-  /// {@template showReactionPickerIndicator}
-  /// Used in [StreamMessageReactionsModal] and [MessageActionsModal]
-  /// {@endtemplate}
-  @Deprecated('Use `showReactionPicker` instead')
-  bool get showReactionPickerIndicator => showReactionPicker;
 
   /// {@template showReactionPickerTail}
   /// Whether or not to show the reaction picker tail
@@ -549,14 +468,13 @@ class StreamMessageWidget extends StatefulWidget {
   final bool showPinHighlight;
 
   /// {@template attachmentBuilders}
-  /// Builder for respective attachment types
+  /// List of attachment builders for rendering attachment widgets pre-defined
+  /// and custom attachment types.
+  ///
+  /// If null, the widget will create a default list of attachment builders
+  /// based on the [Attachment.type] of the attachment.
   /// {@endtemplate}
-  final Map<String, AttachmentBuilder> attachmentBuilders;
-
-  /// {@template customAttachmentBuilders}
-  /// Builder for respective attachment types (user facing builder)
-  /// {@endtemplate}
-  final Map<String, AttachmentBuilder>? customAttachmentBuilders;
+  final List<StreamAttachmentWidgetBuilder>? attachmentBuilders;
 
   /// {@template translateUserAvatar}
   /// Center user avatar with bottom of the message
@@ -586,7 +504,7 @@ class StreamMessageWidget extends StatefulWidget {
   final List<StreamMessageAction> customActions;
 
   /// {@macro onMessageWidgetAttachmentTap}
-  final OnMessageWidgetAttachmentTap? onAttachmentTap;
+  final StreamAttachmentWidgetTapCallback? onAttachmentTap;
 
   /// {@macro attachmentActionsBuilder}
   final AttachmentActionsBuilder? attachmentActionsModalBuilder;
@@ -618,19 +536,7 @@ class StreamMessageWidget extends StatefulWidget {
     Widget Function(BuildContext, Message)? editMessageInputBuilder,
     Widget Function(BuildContext, Message)? textBuilder,
     Widget Function(BuildContext, Message)? quotedMessageBuilder,
-    @Deprecated('''
-    Use [bottomRowBuilderWithDefaultWidget] instead.
-    Will be removed in the next major version.
-    ''') Widget Function(BuildContext, Message)? usernameBuilder,
-    @Deprecated('''
-    Use [bottomRowBuilderWithDefaultWidget] instead.
-    Will be removed in the next major version.
-    ''') BottomRowBuilder? bottomRowBuilder,
     BottomRowBuilderWithDefaultWidget? bottomRowBuilderWithDefaultWidget,
-    @Deprecated('''
-    Use [bottomRowBuilderWithDefaultWidget] instead.
-    Will be removed in the next major version.
-    ''') Widget Function(BuildContext, Message)? deletedBottomRowBuilder,
     void Function(BuildContext, Message)? onMessageActions,
     Message? message,
     StreamMessageThemeData? messageTheme,
@@ -638,9 +544,7 @@ class StreamMessageWidget extends StatefulWidget {
     ShapeBorder? shape,
     ShapeBorder? attachmentShape,
     BorderSide? borderSide,
-    BorderSide? attachmentBorderSide,
     BorderRadiusGeometry? borderRadiusGeometry,
-    BorderRadiusGeometry? attachmentBorderRadiusGeometry,
     EdgeInsetsGeometry? padding,
     EdgeInsets? textPadding,
     EdgeInsetsGeometry? attachmentPadding,
@@ -655,8 +559,6 @@ class StreamMessageWidget extends StatefulWidget {
     void Function(String)? onLinkTap,
     bool? showReactionBrowser,
     bool? showReactionPicker,
-    @Deprecated('Use `showReactionPicker` instead')
-    bool? showReactionPickerIndicator,
     @internal bool? showReactionPickerTail,
     List<Read>? readList,
     ShowMessageCallback? onShowMessage,
@@ -671,7 +573,7 @@ class StreamMessageWidget extends StatefulWidget {
     bool? showFlagButton,
     bool? showPinButton,
     bool? showPinHighlight,
-    Map<String, AttachmentBuilder>? customAttachmentBuilders,
+    List<StreamAttachmentWidgetBuilder>? attachmentBuilders,
     bool? translateUserAvatar,
     OnQuotedMessageTap? onQuotedMessageTap,
     void Function(Message)? onMessageTap,
@@ -685,29 +587,6 @@ class StreamMessageWidget extends StatefulWidget {
     String? imageAttachmentThumbnailCropType,
     AttachmentActionsBuilder? attachmentActionsModalBuilder,
   }) {
-    assert(
-      bottomRowBuilder == null || bottomRowBuilderWithDefaultWidget == null,
-      'You can only use one of the two bottom row builders',
-    );
-
-    var _bottomRowBuilderWithDefaultWidget =
-        bottomRowBuilderWithDefaultWidget ??
-            this.bottomRowBuilderWithDefaultWidget;
-
-    _bottomRowBuilderWithDefaultWidget ??= (context, message, defaultWidget) {
-      final _bottomRowBuilder = bottomRowBuilder ?? this.bottomRowBuilder;
-      if (_bottomRowBuilder != null) {
-        return _bottomRowBuilder(context, message);
-      }
-
-      return defaultWidget.copyWith(
-        onThreadTap: onThreadTap ?? this.onThreadTap,
-        usernameBuilder: usernameBuilder ?? this.usernameBuilder,
-        deletedBottomRowBuilder:
-            deletedBottomRowBuilder ?? this.deletedBottomRowBuilder,
-      );
-    };
-
     return StreamMessageWidget(
       key: key ?? this.key,
       onMentionTap: onMentionTap ?? this.onMentionTap,
@@ -718,7 +597,8 @@ class StreamMessageWidget extends StatefulWidget {
           editMessageInputBuilder ?? this.editMessageInputBuilder,
       textBuilder: textBuilder ?? this.textBuilder,
       quotedMessageBuilder: quotedMessageBuilder ?? this.quotedMessageBuilder,
-      bottomRowBuilderWithDefaultWidget: _bottomRowBuilderWithDefaultWidget,
+      bottomRowBuilderWithDefaultWidget: bottomRowBuilderWithDefaultWidget ??
+          this.bottomRowBuilderWithDefaultWidget,
       onMessageActions: onMessageActions ?? this.onMessageActions,
       message: message ?? this.message,
       messageTheme: messageTheme ?? this.messageTheme,
@@ -726,10 +606,7 @@ class StreamMessageWidget extends StatefulWidget {
       shape: shape ?? this.shape,
       attachmentShape: attachmentShape ?? this.attachmentShape,
       borderSide: borderSide ?? this.borderSide,
-      attachmentBorderSide: attachmentBorderSide ?? this.attachmentBorderSide,
       borderRadiusGeometry: borderRadiusGeometry ?? this.borderRadiusGeometry,
-      attachmentBorderRadiusGeometry:
-          attachmentBorderRadiusGeometry ?? this.attachmentBorderRadiusGeometry,
       padding: padding ?? this.padding,
       textPadding: textPadding ?? this.textPadding,
       attachmentPadding: attachmentPadding ?? this.attachmentPadding,
@@ -743,9 +620,9 @@ class StreamMessageWidget extends StatefulWidget {
           showInChannelIndicator ?? this.showInChannelIndicator,
       onUserAvatarTap: onUserAvatarTap ?? this.onUserAvatarTap,
       onLinkTap: onLinkTap ?? this.onLinkTap,
-      showReactionPicker: showReactionPicker ??
-          showReactionPickerIndicator ??
-          this.showReactionPicker,
+      showReactionPicker: showReactionPicker ?? this.showReactionPicker,
+      showReactionPickerTail:
+          showReactionPickerTail ?? this.showReactionPickerTail,
       onShowMessage: onShowMessage ?? this.onShowMessage,
       showUsername: showUsername ?? this.showUsername,
       showTimestamp: showTimestamp ?? this.showTimestamp,
@@ -759,8 +636,7 @@ class StreamMessageWidget extends StatefulWidget {
       showFlagButton: showFlagButton ?? this.showFlagButton,
       showPinButton: showPinButton ?? this.showPinButton,
       showPinHighlight: showPinHighlight ?? this.showPinHighlight,
-      customAttachmentBuilders:
-          customAttachmentBuilders ?? this.customAttachmentBuilders,
+      attachmentBuilders: attachmentBuilders ?? this.attachmentBuilders,
       translateUserAvatar: translateUserAvatar ?? this.translateUserAvatar,
       onQuotedMessageTap: onQuotedMessageTap ?? this.onQuotedMessageTap,
       onMessageTap: onMessageTap ?? this.onMessageTap,
@@ -817,8 +693,8 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
   /// {@template isGiphy}
   /// `true` if any of the [message]'s attachments are a giphy.
   /// {@endtemplate}
-  bool get isGiphy =>
-      widget.message.attachments.any((element) => element.type == 'giphy');
+  bool get isGiphy => widget.message.attachments
+      .any((element) => element.type == AttachmentType.giphy);
 
   /// {@template isOnlyEmoji}
   /// `true` if [message.text] contains only emoji.
@@ -830,15 +706,14 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
   /// have a [Attachment.titleLink].
   /// {@endtemplate}
   bool get hasNonUrlAttachments => widget.message.attachments
-      .where((it) => it.titleLink == null || it.type == 'giphy')
-      .isNotEmpty;
+      .any((it) => it.type != AttachmentType.urlPreview);
 
   /// {@template hasUrlAttachments}
   /// `true` if any of the [message]'s attachments are a giphy with a
   /// [Attachment.titleLink].
   /// {@endtemplate}
   bool get hasUrlAttachments => widget.message.attachments
-      .any((it) => it.titleLink != null && it.type != 'giphy');
+      .any((it) => it.type == AttachmentType.urlPreview);
 
   /// {@template showBottomRow}
   /// Show the [BottomRow] widget if any of the following are `true`:
@@ -876,7 +751,8 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
   bool get shouldShowEditAction =>
       widget.showEditMessage &&
       !isDeleteFailed &&
-      !widget.message.attachments.any((element) => element.type == 'giphy');
+      !widget.message.attachments
+          .any((element) => element.type == AttachmentType.giphy);
 
   bool get shouldShowResendAction =>
       widget.showResendMessage && (isSendFailed || isUpdateFailed);
@@ -889,7 +765,8 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
   bool get shouldShowEditMessage =>
       widget.showEditMessage &&
       !isDeleteFailed &&
-      !widget.message.attachments.any((element) => element.type == 'giphy');
+      !widget.message.attachments
+          .any((element) => element.type == AttachmentType.giphy);
 
   bool get shouldShowThreadReplyAction =>
       widget.showThreadReplyMessage &&
@@ -961,23 +838,6 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
                       : Alignment.centerLeft,
                   widthFactor: widget.widthFactor,
                   child: Builder(builder: (context) {
-                    var _bottomRowBuilderWithDefaultWidget =
-                        widget.bottomRowBuilderWithDefaultWidget;
-
-                    _bottomRowBuilderWithDefaultWidget ??=
-                        (context, message, defaultWidget) {
-                      final _bottomRowBuilder = widget.bottomRowBuilder;
-                      if (_bottomRowBuilder != null) {
-                        return _bottomRowBuilder(context, message);
-                      }
-
-                      return defaultWidget.copyWith(
-                        onThreadTap: widget.onThreadTap,
-                        usernameBuilder: widget.usernameBuilder,
-                        deletedBottomRowBuilder: widget.deletedBottomRowBuilder,
-                      );
-                    };
-
                     return MessageWidgetContent(
                       streamChatTheme: _streamChatTheme,
                       showUsername: showUsername,
@@ -996,6 +856,12 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
                       textPadding: widget.textPadding,
                       attachmentBuilders: widget.attachmentBuilders,
                       attachmentPadding: widget.attachmentPadding,
+                      attachmentShape: widget.attachmentShape,
+                      onAttachmentTap: widget.onAttachmentTap,
+                      onReplyTap: widget.onReplyTap,
+                      onShowMessage: widget.onShowMessage,
+                      attachmentActionsModalBuilder:
+                          widget.attachmentActionsModalBuilder,
                       avatarWidth: avatarWidth,
                       bottomRowPadding: bottomRowPadding,
                       isFailedState: isFailedState,
@@ -1023,7 +889,7 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
                       onMentionTap: widget.onMentionTap,
                       onQuotedMessageTap: widget.onQuotedMessageTap,
                       bottomRowBuilderWithDefaultWidget:
-                          _bottomRowBuilderWithDefaultWidget,
+                          widget.bottomRowBuilderWithDefaultWidget,
                       onUserAvatarTap: widget.onUserAvatarTap,
                       userAvatarBuilder: widget.userAvatarBuilder,
                     );
@@ -1295,6 +1161,7 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
             showResendMessage: shouldShowResendAction,
             showCopyMessage: shouldShowCopyAction,
             showEditMessage: shouldShowEditAction,
+            showReactionPicker: widget.showReactionPicker,
             showReplyMessage: shouldShowReplyAction,
             showThreadReplyMessage: shouldShowThreadReplyAction,
             showFlagButton: widget.showFlagButton,
