@@ -59,29 +59,6 @@ bool getEffectiveCenterTitle(
   }
 }
 
-/// Shows confirmation dialog
-@Deprecated(
-  '''
-  showConfirmationDialog is deprecated.
-  Use showConfirmationBottomSheet instead.''',
-)
-Future<bool?> showConfirmationDialog(
-  BuildContext context, {
-  required String title,
-  required String okText,
-  Widget? icon,
-  String? question,
-  String? cancelText,
-}) =>
-    showConfirmationBottomSheet(
-      context,
-      title: title,
-      okText: okText,
-      icon: icon,
-      question: question,
-      cancelText: cancelText,
-    );
-
 /// Shows confirmation bottom sheet
 Future<bool?> showConfirmationBottomSheet(
   BuildContext context, {
@@ -168,29 +145,6 @@ Future<bool?> showConfirmationBottomSheet(
     },
   );
 }
-
-/// Shows info dialog
-@Deprecated(
-  '''
-  showInfoDialog is deprecated.
-  Use showInfoBottomSheet instead.''',
-)
-Future<bool?> showInfoDialog(
-  BuildContext context, {
-  required String title,
-  required String okText,
-  Widget? icon,
-  String? details,
-  StreamChatThemeData? theme,
-}) =>
-    showInfoBottomSheet(
-      context,
-      title: title,
-      okText: okText,
-      icon: icon,
-      details: details,
-      theme: theme,
-    );
 
 /// Shows info bottom sheet
 Future<bool?> showInfoBottomSheet(
@@ -376,7 +330,7 @@ String fileSize(dynamic size, [int round = 2]) {
 }
 
 ///
-StreamSvgIcon getFileTypeImage(String? mimeType) {
+StreamSvgIcon getFileTypeImage([String? mimeType]) {
   final subtype = mimeType?.split('/').last;
   switch (subtype) {
     case '7z':
@@ -419,38 +373,19 @@ StreamSvgIcon getFileTypeImage(String? mimeType) {
 }
 
 /// Wraps attachment widget with custom shape
-@Deprecated(
-  '''
-wrapAttachmentWidget is deprecated.
-Use WrapAttachmentWidget instead
-''',
-)
-Widget wrapAttachmentWidget(
-  BuildContext context,
-  Widget attachmentWidget,
-  ShapeBorder attachmentShape,
-  // ignore: avoid_positional_boolean_parameters
-  bool reverse,
-) =>
-    WrapAttachmentWidget(
-      attachmentWidget: attachmentWidget,
-      attachmentShape: attachmentShape,
-    );
-
-/// Wraps attachment widget with custom shape
 class WrapAttachmentWidget extends StatelessWidget {
   /// Builds a [WrapAttachmentWidget].
   const WrapAttachmentWidget({
     super.key,
     required this.attachmentWidget,
-    required this.attachmentShape,
+    this.attachmentShape,
   });
 
   /// The widget to wrap
   final Widget attachmentWidget;
 
   /// The shape of the wrapper
-  final ShapeBorder attachmentShape;
+  final ShapeBorder? attachmentShape;
 
   @override
   Widget build(BuildContext context) {
@@ -497,22 +432,6 @@ int levenshtein(String s, String t, {bool caseSensitive = true}) {
   }
 
   return v1[t.length];
-}
-
-/// An easy way to handle attachment related operations on a message
-extension AttachmentPackagesX on Message {
-  /// This extension will return a List of type [StreamAttachmentPackage] from
-  /// the existing attachments of the message
-  List<StreamAttachmentPackage> getAttachmentPackageList() {
-    final _attachmentPackages = List<StreamAttachmentPackage>.generate(
-      attachments.length,
-      (index) => StreamAttachmentPackage(
-        attachment: attachments[index],
-        message: this,
-      ),
-    );
-    return _attachmentPackages;
-  }
 }
 
 /// PortalLabel that refers to [StreamMessageListView]
