@@ -1,7 +1,7 @@
-export 'attachment_error.dart';
 export 'attachment_upload_state_builder.dart';
-export 'attachment_widget.dart' show AttachmentSource;
 export 'file_attachment.dart';
+export 'gallery_attachment.dart';
 export 'giphy_attachment.dart';
 export 'image_attachment.dart';
+export 'url_attachment.dart';
 export 'video_attachment.dart';
