@@ -1,14 +1,12 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:chewie/chewie.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:photo_view/photo_view.dart';
-import 'package:shimmer/shimmer.dart';
-import 'package:stream_chat_flutter/platform_widget_builder/platform_widget_builder.dart';
+import 'package:stream_chat_flutter/src/attachment/thumbnail/media_attachment_thumbnail.dart';
 import 'package:stream_chat_flutter/src/fullscreen_media/full_screen_media_widget.dart';
+import 'package:stream_chat_flutter/src/fullscreen_media/gallery_navigation_item.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 import 'package:video_player/video_player.dart';
 
@@ -71,7 +69,7 @@ class _FullScreenMediaState extends State<StreamFullScreenMedia> {
     _pageController = PageController(initialPage: widget.startIndex);
     for (var i = 0; i < widget.mediaAttachmentPackages.length; i++) {
       final attachment = widget.mediaAttachmentPackages[i].attachment;
-      if (attachment.type != 'video') continue;
+      if (attachment.type != AttachmentType.video) continue;
       final package = VideoPackage(attachment, showControls: true);
       videoPackages[attachment.id] = package;
     }
@@ -90,7 +88,8 @@ class _FullScreenMediaState extends State<StreamFullScreenMedia> {
       (it) => it.initialize(),
     ));
 
-    if (widget.autoplayVideos && currentAttachment.type == 'video') {
+    if (widget.autoplayVideos &&
+        currentAttachment.type == AttachmentType.video) {
       final package = videoPackages.values
           .firstWhere((e) => e._attachment == currentAttachment);
       package._chewieController?.play();
@@ -270,7 +269,7 @@ class _FullScreenMediaState extends State<StreamFullScreenMedia> {
                   }
                 }
                 if (widget.autoplayVideos &&
-                    currentAttachment.type == 'video') {
+                    currentAttachment.type == AttachmentType.video) {
                   final controller = videoPackages[currentAttachment.id]!;
                   controller._chewieController?.play();
                 }
@@ -289,44 +288,21 @@ class _FullScreenMediaState extends State<StreamFullScreenMedia> {
                           : Colors.black,
                       child: Builder(
                         builder: (context) {
-                          if (attachment.type == 'image' ||
-                              attachment.type == 'giphy') {
-                            final imageUrl = attachment.imageUrl ??
-                                attachment.assetUrl ??
-                                attachment.thumbUrl;
-
-                            return PhotoView(
-                              imageProvider: (imageUrl == null &&
-                                      attachment.localUri != null &&
-                                      attachment.file?.bytes != null)
-                                  ? Image.memory(attachment.file!.bytes!).image
-                                  : CachedNetworkImageProvider(imageUrl!),
-                              errorBuilder: (_, __, ___) =>
-                                  const AttachmentError(),
-                              loadingBuilder: (context, _) {
-                                final image = Image.asset(
-                                  'images/placeholder.png',
-                                  fit: BoxFit.cover,
-                                  package: 'stream_chat_flutter',
-                                );
-                                final colorTheme =
-                                    StreamChatTheme.of(context).colorTheme;
-                                return Shimmer.fromColors(
-                                  baseColor: colorTheme.disabled,
-                                  highlightColor: colorTheme.inputBg,
-                                  child: image,
-                                );
-                              },
+                          if (attachment.type == AttachmentType.image ||
+                              attachment.type == AttachmentType.giphy) {
+                            return PhotoView.customChild(
                               maxScale: PhotoViewComputedScale.covered,
                               minScale: PhotoViewComputedScale.contained,
-                              heroAttributes: PhotoViewHeroAttributes(
-                                tag: widget.mediaAttachmentPackages,
-                              ),
                               backgroundDecoration: const BoxDecoration(
                                 color: Colors.transparent,
                               ),
+                              child: StreamMediaAttachmentThumbnail(
+                                media: attachment,
+                                width: double.infinity,
+                                height: double.infinity,
+                              ),
                             );
-                          } else if (attachment.type == 'video') {
+                          } else if (attachment.type == AttachmentType.video) {
                             final controller = videoPackages[attachment.id]!;
                             if (!controller.initialized) {
                               return const Center(
@@ -357,74 +333,6 @@ class _FullScreenMediaState extends State<StreamFullScreenMedia> {
                   },
                 );
               },
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// A widget for desktop and web users to be able to navigate left and right
-/// through a gallery of images.
-class GalleryNavigationItem extends StatelessWidget {
-  /// Builds a [GalleryNavigationItem].
-  const GalleryNavigationItem({
-    super.key,
-    required this.icon,
-    this.iconSize = 48,
-    required this.onPressed,
-    required this.opacityAnimation,
-    this.left,
-    this.right,
-  });
-
-  /// The icon to display.
-  final Widget icon;
-
-  /// The size of the icon.
-  ///
-  /// Defaults to 48.
-  final double iconSize;
-
-  /// The callback to perform when the button is clicked.
-  final VoidCallback onPressed;
-
-  /// The animation for showing & hiding this widget.
-  final ValueListenable<bool> opacityAnimation;
-
-  /// The left-hand placement of the button.
-  final double? left;
-
-  /// The right-hand placement of the button.
-  final double? right;
-
-  @override
-  Widget build(BuildContext context) {
-    return PlatformWidgetBuilder(
-      desktop: (_, child) => child,
-      web: (_, child) => child,
-      child: Positioned(
-        left: left,
-        right: right,
-        top: MediaQuery.of(context).size.height / 2,
-        child: ValueListenableBuilder<bool>(
-          valueListenable: opacityAnimation,
-          builder: (context, shouldShow, child) {
-            return AnimatedOpacity(
-              opacity: shouldShow ? 1 : 0,
-              duration: kThemeAnimationDuration,
-              child: child,
-            );
-          },
-          child: Material(
-            color: Colors.transparent,
-            type: MaterialType.circle,
-            clipBehavior: Clip.antiAlias,
-            child: IconButton(
-              icon: icon,
-              iconSize: iconSize,
-              onPressed: onPressed,
             ),
           ),
         ),
