@@ -5,6 +5,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:share_plus/share_plus.dart';
+import 'package:stream_chat_flutter/src/attachment/thumbnail/video_attachment_thumbnail.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
 /// {@template streamGalleryFooter}
@@ -94,7 +95,7 @@ class _StreamGalleryFooterState extends State<StreamGalleryFooter> {
                     final url = attachment.imageUrl ??
                         attachment.assetUrl ??
                         attachment.thumbUrl!;
-                    final type = attachment.type == 'image'
+                    final type = attachment.type == AttachmentType.image
                         ? 'jpg'
                         : url.split('?').first.split('.').last;
                     final request = await HttpClient().getUrl(Uri.parse(url));
@@ -217,16 +218,15 @@ class _StreamGalleryFooterState extends State<StreamGalleryFooter> {
                         widget.mediaAttachmentPackages[index];
                     final attachment = attachmentPackage.attachment;
                     final message = attachmentPackage.message;
-                    if (attachment.type == 'video') {
+                    if (attachment.type == AttachmentType.video) {
                       media = MouseRegion(
                         cursor: SystemMouseCursors.click,
                         child: GestureDetector(
                           onTap: () => widget.mediaSelectedCallBack!(index),
                           child: AspectRatio(
                             aspectRatio: 1,
-                            child: StreamVideoThumbnailImage(
-                              video:
-                                  attachment.file?.path ?? attachment.assetUrl,
+                            child: StreamVideoAttachmentThumbnail(
+                              video: attachment,
                             ),
                           ),
                         ),
