@@ -81,10 +81,7 @@ abstract class Translations {
 
   /// The text for showing the unread messages count
   /// in the [StreamMessageListView]
-  String unreadMessagesSeparatorText(
-    @Deprecated('unreadCount is not used anymore and will be removed ')
-    int unreadCount,
-  );
+  String unreadMessagesSeparatorText();
 
   /// The label for "connected" in [StreamConnectionStatusBuilder]
   String get connectedLabel;
@@ -802,7 +799,7 @@ Attachment limit exceeded: it's not possible to add more than $limit attachments
   String get linkDisabledError => 'Links are disabled';
 
   @override
-  String unreadMessagesSeparatorText(int unreadCount) => 'New messages';
+  String unreadMessagesSeparatorText() => 'New messages';
 
   @override
   String get enableFileAccessMessage => 'Please enable access to files'
