@@ -1,5 +1,3 @@
-// ignore_for_file: deprecated_member_use_from_same_package
-
 import 'dart:async';
 import 'dart:math';
 
@@ -110,8 +108,6 @@ class StreamMessageInput extends StatefulWidget {
     this.mediaAttachmentListBuilder,
     this.fileAttachmentBuilder,
     this.mediaAttachmentBuilder,
-    @Deprecated('Use `mediaAttachmentBuilder` instead.')
-    this.attachmentThumbnailBuilders,
     this.focusNode,
     this.sendButtonLocation = SendButtonLocation.outside,
     this.autofocus = false,
@@ -234,10 +230,6 @@ class StreamMessageInput extends StatefulWidget {
 
   /// Builder used to build the media attachment item.
   final AttachmentItemBuilder? mediaAttachmentBuilder;
-
-  /// Map that defines a thumbnail builder for an attachment type.
-  @Deprecated('Use `mediaAttachmentBuilder` instead.')
-  final Map<String, AttachmentThumbnailBuilder>? attachmentThumbnailBuilders;
 
   /// Map that defines a thumbnail builder for an attachment type.
   ///
@@ -1178,7 +1170,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
     }
 
     final containsUrl = quotedMessage.attachments.any((it) {
-      return it.titleLink != null;
+      return it.type == AttachmentType.urlPreview;
     });
 
     return StreamQuotedMessageWidget(
@@ -1221,44 +1213,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
         fileAttachmentListBuilder: widget.fileAttachmentListBuilder,
         mediaAttachmentListBuilder: widget.mediaAttachmentListBuilder,
         fileAttachmentBuilder: widget.fileAttachmentBuilder,
-        mediaAttachmentBuilder: widget.mediaAttachmentBuilder ??
-            // For backward compatibility.
-            // TODO: Remove in the next major release.
-            (context, attachment, onRemovePressed) {
-              final Widget mediaAttachmentThumbnail;
-
-              final builder =
-                  widget.attachmentThumbnailBuilders?[attachment.type];
-              if (builder != null) {
-                mediaAttachmentThumbnail = builder(context, attachment);
-              } else {
-                mediaAttachmentThumbnail = MessageInputMediaAttachmentThumbnail(
-                  attachment: attachment,
-                );
-              }
-
-              return ClipRRect(
-                key: Key(attachment.id),
-                borderRadius: BorderRadius.circular(10),
-                child: Stack(
-                  children: <Widget>[
-                    AspectRatio(
-                      aspectRatio: 1,
-                      child: mediaAttachmentThumbnail,
-                    ),
-                    Positioned(
-                      top: 8,
-                      right: 8,
-                      child: RemoveAttachmentButton(
-                        onPressed: onRemovePressed != null
-                            ? () => onRemovePressed(attachment)
-                            : null,
-                      ),
-                    ),
-                  ],
-                ),
-              );
-            },
+        mediaAttachmentBuilder: widget.mediaAttachmentBuilder,
       ),
     );
   }
@@ -1362,8 +1317,6 @@ class StreamMessageInputState extends State<StreamMessageInput>
       message = message.copyWith(text: '/${message.command} ${message.text}');
     }
 
-    final skipEnrichUrl = _effectiveController.ogAttachment == null;
-
     var shouldKeepFocus = widget.shouldKeepFocusAfterMessage;
     shouldKeepFocus ??= !_commandEnabled;
 
@@ -1387,10 +1340,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
       await WidgetsBinding.instance.endOfFrame;
     }
 
-    await _sendOrUpdateMessage(
-      message: message,
-      skipEnrichUrl: skipEnrichUrl,
-    );
+    await _sendOrUpdateMessage(message: message);
 
     if (mounted) {
       if (shouldKeepFocus) {
@@ -1403,36 +1353,29 @@ class StreamMessageInputState extends State<StreamMessageInput>
 
   Future<void> _sendOrUpdateMessage({
     required Message message,
-    bool skipEnrichUrl = false,
   }) async {
     final channel = StreamChannel.of(context).channel;
 
     try {
       Future sendingFuture;
       if (_isEditing) {
-        sendingFuture = channel.updateMessage(
-          message,
-          skipEnrichUrl: skipEnrichUrl,
-        );
+        sendingFuture = channel.updateMessage(message);
       } else {
-        sendingFuture = channel.sendMessage(
-          message,
-          skipEnrichUrl: skipEnrichUrl,
-        );
+        sendingFuture = channel.sendMessage(message);
       }
 
       final resp = await sendingFuture;
-      if (resp.message?.type == 'error') {
+      if (resp.message?.isError ?? false) {
         _effectiveController.message = message;
       }
       _startSlowMode();
       widget.onMessageSent?.call(resp.message);
     } catch (e, stk) {
       if (widget.onError != null) {
-        widget.onError?.call(e, stk);
-      } else {
-        rethrow;
+        return widget.onError?.call(e, stk);
       }
+
+      rethrow;
     }
   }
 
