@@ -12,18 +12,18 @@ class QuotedMessage extends StatelessWidget {
   const QuotedMessage({
     super.key,
     required this.message,
-    required this.reverse,
     required this.hasNonUrlAttachments,
+    this.textBuilder,
   });
 
   /// {@macro message}
   final Message message;
 
-  /// {@macro reverse}
-  final bool reverse;
-
   /// {@macro hasNonUrlAttachments}
   final bool hasNonUrlAttachments;
+
+  /// {@macro textBuilder}
+  final Widget Function(BuildContext, Message)? textBuilder;
 
   @override
   Widget build(BuildContext context) {
@@ -31,12 +31,15 @@ class QuotedMessage extends StatelessWidget {
     final chatThemeData = StreamChatTheme.of(context);
 
     final isMyMessage = message.user?.id == streamChat.currentUser?.id;
+    final isMyQuotedMessage =
+        message.quotedMessage?.user?.id == streamChat.currentUser?.id;
     return StreamQuotedMessageWidget(
       message: message.quotedMessage!,
       messageTheme: isMyMessage
           ? chatThemeData.otherMessageTheme
           : chatThemeData.ownMessageTheme,
-      reverse: reverse,
+      reverse: !isMyQuotedMessage,
+      textBuilder: textBuilder,
       padding: EdgeInsets.only(
         right: 8,
         left: 8,
