@@ -11,7 +11,7 @@ class StreamColorTheme {
     this.disabled = const Color(0xffdbdbdb),
     this.borders = const Color(0xffecebeb),
     this.inputBg = const Color(0xfff2f2f2),
-    this.appBg = const Color(0xfffcfcfc),
+    this.appBg = const Color(0xfff7f7f8),
     this.barsBg = const Color(0xffffffff),
     this.linkBg = const Color(0xffe9f2ff),
     this.accentPrimary = const Color(0xff005FFF),
@@ -63,8 +63,8 @@ class StreamColorTheme {
     this.disabled = const Color(0xff2d2f2f),
     this.borders = const Color(0xff1c1e22),
     this.inputBg = const Color(0xff13151b),
-    this.appBg = const Color(0xff070A0D),
-    this.barsBg = const Color(0xff101418),
+    this.appBg = const Color(0xff000000),
+    this.barsBg = const Color(0xff121416),
     this.linkBg = const Color(0xff00193D),
     this.accentPrimary = const Color(0xff337eff),
     this.accentError = const Color(0xffFF3742),
