@@ -6,10 +6,9 @@ export 'package:stream_chat_flutter/src/message_widget/quoted_message.dart';
 export 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 
 export 'src/attachment/attachment.dart';
-export 'src/attachment/attachment_title.dart';
+export 'src/attachment/gallery_attachment.dart';
 export 'src/attachment/handler/stream_attachment_handler.dart';
 export 'src/attachment/image_attachment.dart';
-export 'src/attachment/image_group.dart';
 export 'src/attachment/stream_attachment_package.dart';
 export 'src/attachment/url_attachment.dart';
 export 'src/attachment/video_attachment.dart';
@@ -26,11 +25,9 @@ export 'src/channel/channel_header.dart';
 export 'src/channel/channel_info.dart';
 export 'src/channel/channel_list_header.dart';
 export 'src/channel/channel_name.dart';
-export 'src/channel/channel_preview.dart';
 export 'src/channel/stream_channel_avatar.dart';
 export 'src/channel/stream_channel_name.dart';
 export 'src/channel/stream_message_preview_text.dart';
-export 'src/fullscreen_media/fsm_enums.dart';
 export 'src/fullscreen_media/full_screen_media.dart';
 export 'src/fullscreen_media/full_screen_media_builder.dart';
 export 'src/gallery/gallery_footer.dart';
@@ -93,10 +90,10 @@ export 'src/stream_chat.dart';
 export 'src/stream_chat_configuration.dart';
 export 'src/theme/stream_chat_theme.dart';
 export 'src/theme/themes.dart';
-export 'src/user/user_item.dart';
 export 'src/user/user_mention_tile.dart';
 export 'src/utils/device_segmentation.dart';
 export 'src/utils/extensions.dart';
 export 'src/utils/helpers.dart';
 export 'src/utils/typedefs.dart';
+// TODO: Remove this in favor of StreamVideoAttachmentThumbnail.
 export 'src/video/video_thumbnail_image.dart';
