@@ -2,6 +2,7 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:stream_chat_flutter/src/attachment/thumbnail/image_attachment_thumbnail.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
 import '../mocks.dart';
@@ -18,6 +19,27 @@ void main() {
       final themeData = ThemeData();
       final streamTheme = StreamChatThemeData.fromTheme(themeData);
 
+      final attachments = [
+        Attachment(
+          type: 'image',
+          title: 'example.png',
+          imageUrl:
+              'https://logowik.com/content/uploads/images/flutter5786.jpg',
+          extraData: const {
+            'mime_type': 'png',
+          },
+        ),
+        Attachment(
+          type: 'image',
+          title: 'example.png',
+          imageUrl:
+              'https://logowik.com/content/uploads/images/flutter5786.jpg',
+          extraData: const {
+            'mime_type': 'png',
+          },
+        ),
+      ];
+
       await tester.pumpWidget(
         MaterialApp(
           home: StreamChatTheme(
@@ -25,33 +47,23 @@ void main() {
             child: StreamChannel(
               channel: channel,
               child: SizedBox(
-                child: StreamImageGroup(
-                  messageTheme: streamTheme.ownMessageTheme,
+                child: StreamGalleryAttachment(
                   constraints: BoxConstraints.tight(const Size(
                     300,
                     300,
                   )),
                   message: Message(),
-                  images: [
-                    Attachment(
-                      type: 'image',
-                      title: 'example.png',
-                      imageUrl:
-                          'https://logowik.com/content/uploads/images/flutter5786.jpg',
-                      extraData: const {
-                        'mime_type': 'png',
-                      },
-                    ),
-                    Attachment(
-                      type: 'image',
-                      title: 'example.png',
-                      imageUrl:
-                          'https://logowik.com/content/uploads/images/flutter5786.jpg',
-                      extraData: const {
-                        'mime_type': 'png',
-                      },
-                    ),
-                  ],
+                  attachments: attachments,
+                  itemBuilder: (context, index) {
+                    final attachment = attachments[index];
+
+                    return StreamImageAttachmentThumbnail(
+                      image: attachment,
+                      width: double.infinity,
+                      height: double.infinity,
+                      fit: BoxFit.cover,
+                    );
+                  },
                 ),
               ),
             ),
