@@ -1,11 +1,32 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:stream_chat_flutter/scrollable_positioned_list/scrollable_positioned_list.dart';
 import 'package:stream_chat_flutter/src/message_list_view/floating_date_divider.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+import '../mocks.dart';
+
 void main() {
   testWidgets('FloatingDateDivider', (tester) async {
+    final client = MockClient();
+    final clientState = MockClientState();
+
+    when(() => client.state).thenReturn(clientState);
+    when(() => clientState.currentUser).thenReturn(OwnUser(id: 'user-id'));
+
+    final createdAt = DateTime.now();
+
+    final itemPositionsListener = ValueNotifier(
+      [
+        const ItemPosition(
+          index: 0,
+          itemLeadingEdge: 0,
+          itemTrailingEdge: 0,
+        ),
+      ],
+    );
+
     await tester.pumpWidget(
       MaterialApp(
         home: Scaffold(
@@ -14,12 +35,9 @@ void main() {
               FloatingDateDivider(
                 reverse: false,
                 itemCount: 3,
-                itemPositionListener: ItemPositionsListener.create(),
-                messages: [
-                  Message(),
-                  Message(),
-                  Message(),
-                ],
+                itemPositionListener: itemPositionsListener,
+                messages: [Message(createdAt: createdAt)],
+                dateDividerBuilder: (dateTime) => Text('$dateTime'),
               ),
             ],
           ),
@@ -27,6 +45,6 @@ void main() {
       ),
     );
 
-    expect(find.byType(Positioned), findsOneWidget);
+    expect(find.text('$createdAt'), findsOneWidget);
   });
 }
