@@ -421,7 +421,7 @@ class StreamChatLocalizationsKo extends GlobalStreamChatLocalizations {
   String get linkDisabledError => '링크가 비활성화되었습니다.';
 
   @override
-  String unreadMessagesSeparatorText(int unreadCount) => '새 메시지.';
+  String unreadMessagesSeparatorText() => '새 메시지.';
 
   @override
   String get enableFileAccessMessage => '친구와 공유할 수 있도록 파일에 대한 액세스를 허용하세요.';
